+++ CHANGELOG.md
@@ -1,11 +1,33 @@
-## 1.0.2+2
+Date format: DD/MM/YYYY
+
+## 2.0.0 - [15/02/2021]
+
+- **BREAKING**: Removed `completeInfoBook` (it was redundant)
+- **NEW**: `BookInfo.toString()`
+- Performance improvements
+- Formatted files
+- Upgraded depedencies
+
+## 1.0.2+4 - [28/08/2021]
+
+- Hotfix startIndex and maxResults conflict issue
+
+## 1.0.2+3 - [28/08/2021]
+
+- hotfix thumbnail in a map
+
+## 1.0.2+2 - [28/08/2021]
+
 - hotfix book not found
 
-## 1.0.2+1
+## 1.0.2+1 - [28/08/2021]
+
 - Hotfix startIndex issue
 
-## 1.0.2
+## 1.0.2 - [27/08/2021]
+
 - Export the book object
 
-## 1.0.1
+## 1.0.1 - [27/08/2021]
+
 - First version